VISUAL STUDIO CODE
This is the final version of this project.

In the final version,  you'll observe that if the user input 

-  The Wrong Information 
-  Wrong format,

The user will stuck in the while loop and will get the exception in the form of the output. 

In this you'll get same output as the first version of this project, but you can observe the difference between the scripts

First Version - Only simple input and got the result.

Final Version - Using Loops. If-Elif-Else, Try-Exception, & Functions .